+++ src/contexts/AuthUIContextProvider/provider.tsx
@@ -34,10 +34,13 @@ export const AuthUIContextProvider: React.FC<AuthUIContextProviderProps> = (prop
             loginHeader: props.loginHeader,
             passwordRequirements: props.passwordRequirements,
             customAccountDetails: props.customAccountDetails,
+            registrationSuccessScreen: props.registrationSuccessScreen,
+            accountAlreadyExistsScreen: props.accountAlreadyExistsScreen,
         };
 
         return propsForContext;
     }, [
+        props.accountAlreadyExistsScreen,
         props.allowDebugMode,
         props.authActions,
         props.background,
@@ -45,6 +48,7 @@ export const AuthUIContextProvider: React.FC<AuthUIContextProviderProps> = (prop
         props.contactPhone,
         props.contactPhoneLink,
         props.customAccountDetails,
+        props.registrationSuccessScreen,
         props.enableInviteRegistration,
         props.enableResetPassword,
         props.htmlEula,