+++ CHANGELOG.md
@@ -5,11 +5,17 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## v3.1.0 (Unreleased)
+
+### Added
+
+-   Updated type definitions to allow for a custom success screen in the Registration flow.
+
 ## v3.0.0 (March 29, 2021)
 
 ### Added
 
--   Updated type definitions to allow for further customization of the elements on the Login screen
+-   Updated type definitions to allow for further customization of the elements on the Login screen.
 -   Updated type definitions to allow for custom forms / data collection in the Registration flow.
 -   Peer dependency support for React 17.